Benchmarks ruby vs python vs golang vs perl
Ruby is a current dependency because env dumping and diffing is way too
hard to do in bash. Dictionary support is only available in it's latest
versions and dumping in a shell-escaped version is not readily
available.

The problem with ruby is that it doesn't come pre-installed on most
POSIX systems and I believe that it thus is a hindrance to direnv's
adoption. I also want direnv to be as easily installable and available
as possible and removing a dependency is always a good thing.

From there we can go two ways:
either use a more widely available runtime like python or perl. Or add a
compile-time dependency like golang.
And when using a runtime, only use the stdlib to avoid introducing another
dependency.

Because direnv is executed every time the prompt is displayed, it
shouldn't be noticeable that it's installed. Otherwise direnv is a
hindrance to the proper usage of the system and I prefer to source the
environments manually (call me excessive ?). So I made a set of
benchmarks in this commit. On my machine, here are the results:

```
$ ./libexec/bench.rb
For 100 iterations:
                                     user     system      total        real

empty iterations                 0.010000   0.040000   0.250000 (  0.274686)

bash   startup                   0.010000   0.060000   0.530000 (  0.542160)
bash   existing env              0.010000   0.050000   1.710000 (  1.913937)
bash   no env                    0.010000   0.050000   0.960000 (  1.011029)

ruby   startup                   0.010000   0.050000   1.540000 (  1.622935)
ruby   dump                      0.010000   0.060000   1.350000 (  1.448645)
ruby   diff + dump               0.010000   0.060000   2.600000 (  2.893097)
ruby   export                    0.010000   0.060000   6.700000 (  7.327523)
ruby   eval + export             0.020000   0.060000   7.740000 (  8.586604)

python startup                   0.010000   0.060000   2.880000 (  3.025044)
python dump                      0.010000   0.070000   3.180000 (  3.311468)
python diff + dump               0.010000   0.060000  11.730000 ( 12.121832)

perl   startup                   0.010000   0.060000   0.930000 (  1.056434)
perl   dump (using Storable)     0.020000   0.080000   8.280000 (  8.558934)
perl   dump (using Dumper)       0.010000   0.070000   6.870000 (  7.053147)

golang dump                      0.010000   0.070000   0.790000 (  0.894930)
golang diff + dump               0.010000   0.070000   1.240000 (  1.395290)
```

= Analysis

The original ruby export when loading an environment takes 80 ms to
execute, which is beyond the user-noticable 100 ms barrier.

Both python and perl times are even slower than ruby. For perl I didn't
implement the full solution since dumping was way too slow already (and
the language is a little to weird for me).

golang is a clear winner in terms of speed and could help keep the total
execution time below the 100 ms even on slower systems. Replacing the
whole direnv dispatch would also avoid lots of syscalls and fork-exec.